requirements.txt
fastapi
uvicorn
sentence-transformers
pinecone
google-generativeai
